+++ backend/src/main/java/at/htl/leonding/features/projectImage/ProjectImageResource.java
@@ -1,12 +1,15 @@
 package at.htl.leonding.features.projectImage;
 
+import at.htl.leonding.features.security.CustomPrincipal;
 import jakarta.annotation.security.PermitAll;
 import jakarta.annotation.security.RolesAllowed;
 import jakarta.inject.Inject;
 import jakarta.transaction.Transactional;
 import jakarta.ws.rs.*;
+import jakarta.ws.rs.core.Context;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
+import jakarta.ws.rs.core.SecurityContext;
 
 @Path("/project-images")
 @Produces(MediaType.APPLICATION_JSON)
@@ -33,7 +36,13 @@ public class ProjectImageResource {
     @Path("/{id}")
     @RolesAllowed({"admin"})
     @Transactional
-    public Response deleteImageById(@PathParam("id") long id) {
+    public Response deleteImageById(@PathParam("id") long id, @Context SecurityContext ctx) {
+
+        CustomPrincipal p = (CustomPrincipal) ctx.getUserPrincipal();
+        if (!p.isAdmin()) {
+            return Response.status(Response.Status.UNAUTHORIZED).build();
+        }
+
         projectImageRepository.deleteById(id);
 
         return Response.noContent().build();
